Delete `SocketChessServer.java`.
`SocketChessServer` was useful for debugging the `Decider`s before
the JavaScript + WebSocket client came into being.  But, it's
pretty bad code, and probably not worht saving.

Time that could be spent refactoring and adding featurs would better be
spent refactoring and adding features to `StandardStreamChessServer`
(soon to be called `StreamChessServer`), and the client.