Make ant test target more specific.
Previously, the test target looked for all files including the string "Test"
in the src directory, and tried to run them all.

Now, it only runs the `AllTests` test suite.  This is good,
because it allows us to explicitly define which files are in the test suite,
and also to use test suites as a way to organize which tests are run.  For
example, we could envision a future where we have separate suites for
unit tests and system tests, and we want to specify different ant targets
for each suite.