Seriously refactor `ChessMove`.
Previously `ChessMove` was a class, and `PromotionMove` and `CastlingMove`
were subclasses of it.

Now, `ChessMove` is an interface, and all three of `NormalChessMove`,
`PromotionMove`, and `CastlingMove` implement that interface.

`PromotionMove` and `CastlingMove` no longer extend the `NormalChessMove`
class (which used to be the `ChessMove` class).  Instead, they
use `NormalChessMove` by composition.  This change was mostly executed
so that we wouldn't run into weird subclass-equality issues.
(This change should have happened in a separate commit.  I got lazy,
and didn't separate the changes.  Sorry.)