Add `equals` and `hashCode` methods to `CastlingMove` and `PromotionMove`.
These will be useful fro implementing tests for `ChessPosition` and
`ChessMove` implementations.

Tests for these methods will be added in an upcoming commit.  (I want
to write the `AbstractChessPosition` tests before I write the tests
for the `ChessMove` implementations, but I needed these equality
methods in order to test `AbstractChessPosition`.  This is a result
of the cyclic dependence between `ChessPosition` and `ChessMove`.
Someday I'll see if I can eliminate that.)