Remove `ChessMove.getDelta`.
It doesn't add any info that's not already conveyed by `getStart` and `getEnd`.
Its existence (probably) led to a slight efficiency gain. But, we're not looking
for efficiency right now-- we're looking for clear code.

If we want that method to come back, ever, it should probably come back
as `EfficientChessMove.getDelta`... where `EfficientChessMove` extends
`ChessMove`.  (That name, `EfficientChessMove` could probably be improved.)