Fix bug causing certain pawn checks not to be recognized.
Previously, `AbstractChessPosition.checked` determined if a king
was in check by seeing whether any `NormalChessMove`s could be made
to capture the king.  But, it ignored `PromotionMove`s.  So,
if a pawn on the seventh rank checked a king on the eight rank,
it would not be detected that the king was checked!

Now, it is fixed.