Make `AlphaBeta` and `Minimax` track current depth.
This allows us to get accurate values for `TerminalScore`s'
`pliesUntilEnd` attributes, and in a way that is convenient for
alpha-beta pruning.

While I did this work, it became clear to me that `Heuristic.terminalValue`
didn't belong in `Heuristic`.  So, it was deleted.  That was the only method
implemented in`AbstractHeuristic`, so `AbstractHeuristic` was deleted.
(I haven't found the right place for that code to go, yet.  Currently,
it has been duplicated in `AlphaBeta` and `Minimax`, but eventually I expect
to factor it out into a better location.